Breaking content importer into smaller parts, refactoring
