+++ concrete/config/app.php
@@ -194,6 +194,58 @@ return array(
         'workflow_type',
     ),
 
+    'importer_routines' => array(
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportGroupsRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportSinglePageStructureRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportStacksStructureRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportBlockTypesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportBlockTypeSetsRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportConversationEditorsRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportConversationRatingTypesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportConversationFlagTypesTypesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportPageTypePublishTargetTypesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportPageTypeComposerControlTypesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportBannedWordsRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportSocialLinksRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportTreesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportFileImportantThumbnailTypesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportFeaturesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportFeatureCategoriesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportGatheringDataSourcesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportGatheringItemTemplateTypesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportGatheringItemTemplatesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportAttributeCategoriesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportAttributeTypesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportWorkflowTypesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportWorkflowProgressCategoriesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportAttributesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportAttributeSetsRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportExpressEntitiesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportExpressAssociationsRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportExpressFormsRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportExpressRelationsRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportThemesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportPermissionKeyCategoriesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportPermissionAccessEntityTypesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportPermissionsRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportJobsRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportJobSetsRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportPageTemplatesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportPageTypesBaseRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportPageStructureRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportPageFeedsRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportPageTypeTargetsRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportPageTypeDefaultsRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportSinglePageContentRoutine',
+
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportStacksContentRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportPageContentRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportPackagesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportConfigValuesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportSystemCaptchaLibrariesRoutine',
+        'Concrete\Core\Backup\ContentImporter\Importer\Routine\ImportSystemContentEditorSnippetsRoutine',
+    ),
+
     /*
      * Core Routes
      */
