+++ concrete/src/Backup/ContentImporter/Importer/Manager.php
@@ -1,0 +1,20 @@
+<?php
+namespace Concrete\Core\Backup\ContentImporter\Importer;
+
+use Concrete\Core\Backup\ContentImporter\Importer\Routine\RoutineInterface;
+
+class Manager
+{
+    protected $routines = array();
+
+    public function registerImporterRoutine(RoutineInterface $routine)
+    {
+        $this->routines[$routine->getHandle()] = $routine;
+    }
+
+    public function getImporterRoutines()
+    {
+        return $this->routines;
+    }
+
+}
