+++ concrete/src/Backup/ContentImporter/Importer/Routine/AbstractRoutine.php
@@ -1,0 +1,19 @@
+<?php
+namespace Concrete\Core\Backup\ContentImporter\Importer\Routine;
+
+abstract class AbstractRoutine implements RoutineInterface
+{
+
+    protected static function getPackageObject($pkgHandle)
+    {
+        $pkg = false;
+        if ($pkgHandle) {
+            $pkgHandle = (string) $pkgHandle;
+            $pkg = \Package::getByHandle($pkgHandle);
+        }
+
+        return $pkg;
+    }
+
+
+}
