+++ concrete/src/Backup/ContentImporter/Importer/Routine/RoutineInterface.php
@@ -1,0 +1,8 @@
+<?php
+namespace Concrete\Core\Backup\ContentImporter\Importer\Routine;
+
+interface RoutineInterface
+{
+    public function getHandle();
+    public function import(\SimpleXMLElement $element);
+}
